# Performance: decrease runtime overhead for constructing HasTraits (up to 20x faster)