This makes creating ipywidgets about 2x faster (ignoring comm construction).

Related to, but superseded https://github.com/ipython/traitlets/pull/639