@maartenbreddels 2022-09-14
The ipywidget tests fail, since it adds `_comm_default` at runtime, https://github.com/jupyter-widgets/ipywidgets/pull/3590 fixes that and is also a better method anyway. The question is, is this a breaking change? I don't think there is any mention of adding `_trait_default` at runtime, and it causes a significant runtime cost.