@maartenbreddels 2022-09-26
I have another commit that gives a good performance win in a situation where a lot of default are 'static' (like a fixed default value for string and ints). This happens a lot for ipywidgets.

Should I push to this branch? or should I wait to get this in first (I could open a new PR and later rebase against main)