@maartenbreddels 2022-09-26
I've been benchmarking this with pytest-benchmark using this script:
```python
from traitlets import HasTraits, Int, Unicode

class SingleTrait(HasTraits):
    foo = Int()


class ManyTrait(HasTraits):
    foo0 = Unicode("foo0").tag(sync=True)
    foo1 = Unicode("foo1").tag(sync=True)
    foo2 = Unicode("foo2").tag(sync=True)
    foo3 = Unicode("foo3").tag(sync=True)
    foo4 = Unicode("foo4").tag(sync=True)
    foo5 = Unicode("foo5").tag(sync=True)
    foo6 = Unicode("foo6").tag(sync=True)
    foo7 = Unicode("foo7").tag(sync=True)
    foo8 = Unicode("foo8").tag(sync=True)
    foo9 = Unicode("foo9").tag(sync=True)



def create(N, cls):
    for i in range(N):
        cls()


def create_and_access(N, cls):
    keys = cls.class_trait_names(sync=True)
    for i in range(N):
        obj = create()
        for key in keys:
            getattr(obj, key)


def test_single_trait(benchmark):
    benchmark(create, cls=SingleTrait, N=1_000)

def test_many_trait(benchmark):
    benchmark(create, cls=ManyTrait, N=1_000)

def test_single_trait_and_access(benchmark):
    benchmark(create_and_access, cls=SingleTrait, N=1_000)

def test_many_trait_and_access(benchmark):
    benchmark(create_and_access, cls=ManyTrait, N=1_000)

```

# Results
## Main a35dd0b686d9cbdfc65ce14e4889ea39980d156d
```
-------------------------------------------------------------------------------------- benchmark: 4 tests --------------------------------------------------------------------------------------
Name (time in ms)                    Min                Max               Mean            StdDev             Median               IQR            Outliers      OPS            Rounds  Iterations
------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
test_single_trait                12.4555 (1.0)      12.9838 (1.0)      12.5927 (1.0)      0.1114 (1.62)     12.5714 (1.0)      0.1253 (2.96)         14;2  79.4108 (1.0)          71           1
test_single_trait_and_access     13.0052 (1.04)     13.5203 (1.04)     13.1818 (1.05)     0.1235 (1.80)     13.1553 (1.05)     0.1789 (4.23)         21;0  75.8624 (0.96)         75           1
test_many_trait                  14.0976 (1.13)     14.4672 (1.11)     14.3162 (1.14)     0.0686 (1.0)      14.3159 (1.14)     0.0423 (1.0)         19;16  69.8510 (0.88)         69           1
test_many_trait_and_access       57.5647 (4.62)     59.3905 (4.57)     58.3111 (4.63)     0.5214 (7.60)     58.1785 (4.63)     0.8110 (19.18)         5;0  17.1494 (0.22)         18           1
------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
```
## before the last commit 166cef4ffca4896528b5079839dfb64ef67d4b1b
```
--------------------------------------------------------------------------------------- benchmark: 4 tests --------------------------------------------------------------------------------------
Name (time in ms)                    Min                Max               Mean            StdDev             Median               IQR            Outliers       OPS            Rounds  Iterations
-------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
test_single_trait                 2.2666 (1.0)       2.5554 (1.0)       2.3552 (1.0)      0.0274 (1.63)      2.3493 (1.0)      0.0152 (1.06)        50;47  424.5879 (1.0)         400           1
test_single_trait_and_access      2.4376 (1.08)      2.6091 (1.02)      2.5242 (1.07)     0.0169 (1.0)       2.5228 (1.07)     0.0151 (1.06)        54;24  396.1647 (0.93)        394           1
test_many_trait                   2.6998 (1.19)      2.9632 (1.16)      2.8028 (1.19)     0.0299 (1.77)      2.7969 (1.19)     0.0143 (1.0)         47;50  356.7877 (0.84)        367           1
test_many_trait_and_access       27.8987 (12.31)    29.0280 (11.36)    28.0583 (11.91)    0.2163 (12.83)    27.9914 (11.91)    0.0954 (6.67)          4;4   35.6400 (0.08)         36           1
-------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
```
## ae8c6cde983a689602f747932f55fa3780a4a1fe
```
------------------------------------------------------------------------------------- benchmark: 4 tests ------------------------------------------------------------------------------------
Name (time in ms)                   Min               Max              Mean            StdDev            Median               IQR            Outliers       OPS            Rounds  Iterations
---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
test_single_trait                2.4483 (1.0)      2.7051 (1.0)      2.5550 (1.0)      0.0328 (1.13)     2.5482 (1.0)      0.0225 (1.44)        59;49  391.3851 (1.0)         370           1
test_single_trait_and_access     2.6786 (1.09)     2.9397 (1.09)     2.7569 (1.08)     0.0308 (1.06)     2.7498 (1.08)     0.0218 (1.39)        52;41  362.7242 (0.93)        362           1
test_many_trait                  3.3650 (1.37)     3.7456 (1.38)     3.4842 (1.36)     0.0291 (1.0)      3.4812 (1.37)     0.0156 (1.0)         22;22  287.0108 (0.73)        278           1
test_many_trait_and_access       5.3269 (2.18)     5.9227 (2.19)     5.5025 (2.15)     0.0671 (2.31)     5.4879 (2.15)     0.0230 (1.47)        13;19  181.7341 (0.46)        170           1
---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
```

Looking at the median, we see a >10x improvement for `HasTrait` with many 'static'+immutable data, which is quite common in ipywidgets (see e.g. the Layout widget https://github.com/jupyter-widgets/ipywidgets/blob/master/python/ipywidgets/ipywidgets/widgets/widget_layout.py )

where the last commit really helps with a factor of `~5x`