@maartenbreddels 2022-09-28
Found another performance issue: We are calling `Descriptor.instance_init` for each HasTraits we create, for each descriptor, while actually many of them don't do anything.
I've refactored it such that only at class construction we collect all `instance_init` methods that do something. We by default assume that we "do something", by adding `instance_init` to a list of functions to be called.
However, many of the TypeTraits opt out of this by overriding `subclass_init` to *not* add it's own `instance_init` to the list of functions.
This makes it almost perfectly backwards compatible, except...
When someone inherits from say `Int`, its `instance_init` will not be called. We could make it fully compatible by checking `if type(self) == Int` (and similar for the other traits). Let me know, I'm happy to add this in.


When creating a 1000 HasTraits with 30 traits that do not need an `instance_init`, we avoid 30000 unneeded function calls. 

## Performance benchmark:
### Before
```
------------------------------------------------------------------------------------- benchmark: 4 tests ------------------------------------------------------------------------------------
Name (time in ms)                   Min               Max              Mean            StdDev            Median               IQR            Outliers       OPS            Rounds  Iterations
---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
test_single_trait                2.2703 (1.0)      2.5766 (1.0)      2.3246 (1.0)      0.0510 (1.50)     2.3051 (1.0)      0.0656 (2.79)        65;13  430.1731 (1.0)         407           1
test_single_trait_and_access     2.2841 (1.01)     2.7480 (1.07)     2.3585 (1.01)     0.0789 (2.31)     2.3257 (1.01)     0.0752 (3.20)        62;36  423.9969 (0.99)        430           1
test_many_trait                  2.8105 (1.24)     3.0434 (1.18)     2.8452 (1.22)     0.0341 (1.0)      2.8343 (1.23)     0.0235 (1.0)         36;30  351.4728 (0.82)        287           1
test_many_trait_and_access       4.6244 (2.04)     4.9061 (1.90)     4.6736 (2.01)     0.0457 (1.34)     4.6592 (2.02)     0.0415 (1.77)        24;15  213.9672 (0.50)        213           1
---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
```

### After
```
------------------------------------------------------------------------------------- benchmark: 4 tests ------------------------------------------------------------------------------------
Name (time in ms)                   Min               Max              Mean            StdDev            Median               IQR            Outliers       OPS            Rounds  Iterations
---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
test_single_trait                2.2464 (1.0)      2.4565 (1.01)     2.2825 (1.00)     0.0252 (1.0)      2.2769 (1.00)     0.0230 (1.32)        58;22  438.1106 (1.00)        422           1
test_many_trait                  2.2525 (1.00)     2.4332 (1.00)     2.2818 (1.0)      0.0257 (1.02)     2.2746 (1.0)      0.0223 (1.28)        55;32  438.2443 (1.0)         436           1
test_single_trait_and_access     2.2594 (1.01)     2.4330 (1.0)      2.2858 (1.00)     0.0254 (1.01)     2.2781 (1.00)     0.0174 (1.0)         42;36  437.4761 (1.00)        427           1
test_many_trait_and_access       4.1723 (1.86)     4.4918 (1.85)     4.2298 (1.85)     0.0608 (2.42)     4.2072 (1.85)     0.0509 (2.92)        33;26  236.4174 (0.54)        230           1
---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------

```

This saves another ~10% for HasTraits with 10 traits. I've also tested this in a more extreme case with 100 traits, which gives a 25% performance increase. This shows this approach scales much better.

Looking at creating a 100 trait object in viztracer, we see:
# before
<img width="362" alt="image" src="https://user-images.githubusercontent.com/1765949/192728287-70af601e-751e-4704-930f-cfe5a4420639.png">

`setup_instance` takes most of the time.
# after
<img width="329" alt="image" src="https://user-images.githubusercontent.com/1765949/192728570-7a0a5698-cd8a-4a5e-8e4d-21f76095dbae.png">

So instead of `setup_instance` taking most of the time creating an object, we only spend time doing a dict copy with the trait values (the only overhead that we cannot avoid). Now the last bottle neck seems to be the HasTraits constructor (more specifically the context manager)